+++ packages/terra-framework-docs/src/terra-dev-site/test/slide-group/AnimatedSlideGroup.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import classNames from 'classnames/bind';
-import SlideGroup from '../../../SlideGroup';
-import styles from '../../doc/example/SlideGroupDemo.module.scss';
+import SlideGroup from 'terra-slide-group';
+import styles from '../../doc/slide-group/example/SlideGroupDemo.module.scss';
 
 const cx = classNames.bind(styles);
 
@@ -41,7 +41,7 @@ class SlideGroupExample extends React.Component {
 
     return (
       <div className={cx('content-wrapper')}>
-        <SlideGroup items={slides} id="SlideGroup" />
+        <SlideGroup items={slides} isAnimated id="SlideGroup" />
       </div>
     );
   }