+++ packages/terra-framework-docs/CHANGELOG.md
@@ -18,6 +18,7 @@
   * Migrate docs from terra-disclosure-manager
   * Migrate docs from terra-embedded-content-consumer
   * Migrate docs from terra-slide-panel.
+  * Migrate docs from terra-slide-group.
 
 ## 1.7.0 - (March 8, 2022)